+++ packages/components/ui/src/icons/IconCheckSquare.svelte
@@ -1,0 +1,21 @@
+<script>
+	import Icon from './Icon.svelte';
+
+	export let size = null;
+	export let fill = null;
+	export let stroke = null;
+	export let strokeWidth = null;
+</script>
+
+<Icon
+	{fill}
+	{stroke}
+	{strokeWidth}
+	height={size}
+	width={size}
+>
+	<polyline points="9 11 12 14 22 4" />
+	<path d="M21 12v7a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h11" />
+</Icon>
+
+<!-- feathericons/feather/icons/check-square.svg -->