+++ packages/components/ui/src/icons/IconChevronDown.svelte
@@ -1,0 +1,20 @@
+<script>
+	import Icon from './Icon.svelte';
+
+	export let size = null;
+	export let fill = null;
+	export let stroke = null;
+	export let strokeWidth = null;
+</script>
+
+<Icon
+	{fill}
+	{stroke}
+	{strokeWidth}
+	height={size}
+	width={size}
+>
+	<polyline points="6 9 12 15 18 9" />
+</Icon>
+
+<!-- feathericons/feather/icons/chevron-down.svg -->