+++ packages/components/ui/src/icons/IconClipboard.svelte
@@ -1,0 +1,21 @@
+<script>
+	import Icon from './Icon.svelte';
+
+	export let size = null;
+	export let fill = null;
+	export let stroke = null;
+	export let strokeWidth = null;
+</script>
+
+<Icon
+	{fill}
+	{stroke}
+	{strokeWidth}
+	height={size}
+	width={size}
+>
+	<path d="M16 4h2a2 2 0 0 1 2 2v14a2 2 0 0 1-2 2H6a2 2 0 0 1-2-2V6a2 2 0 0 1 2-2h2" />
+	<rect x="8" y="2" width="8" height="4" rx="1" ry="1" />
+</Icon>
+
+<!-- feathericons/feather/icons/clipboard.svg -->