+++ packages/components/ui/src/icons/IconExternalLink.svelte
@@ -1,0 +1,22 @@
+<script>
+	import Icon from './Icon.svelte';
+
+	export let size = null;
+	export let fill = null;
+	export let stroke = null;
+	export let strokeWidth = null;
+</script>
+
+<Icon
+	{fill}
+	{stroke}
+	{strokeWidth}
+	height={size}
+	width={size}
+>
+	<path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6" />
+	<polyline points="15 3 21 3 21 9" />
+	<line x1="10" y1="14" x2="21" y2="3" />
+</Icon>
+
+<!-- feathericons/feather/icons/external-link.svg -->