+++ packages/components/ui/src/index.js
@@ -1,1 +1,22 @@
 export {default as Switch} from './Switch.svelte';
+
+// icons
+
+export {default as Icon} from './icons/Icon.svelte';
+export {default as IconCheck} from './icons/IconCheck.svelte';
+export {default as IconCheckSquare} from './icons/IconCheckSquare.svelte';
+export {default as IconChevronDown} from './icons/IconChevronDown.svelte';
+export {default as IconChevronLeft} from './icons/IconChevronLeft.svelte';
+export {default as IconChevronRight} from './icons/IconChevronRight.svelte';
+export {default as IconChevronUp} from './icons/IconChevronUp.svelte';
+export {default as IconClipboard} from './icons/IconClipboard.svelte';
+export {default as IconDelete} from './icons/IconDelete.svelte';
+export {default as IconDownload} from './icons/IconDownload.svelte';
+export {default as IconExternalLink} from './icons/IconExternalLink.svelte';
+export {default as IconFilter} from './icons/IconFilter.svelte';
+export {default as IconGlobe} from './icons/IconGlobe.svelte';
+export {default as IconHelpCircle} from './icons/IconHelpCircle.svelte';
+export {default as IconInfo} from './icons/IconInfo.svelte';
+export {default as IconSettings} from './icons/IconSettings.svelte';
+export {default as IconSliders} from './icons/IconSliders.svelte';
+export {default as IconSquare} from './icons/IconSquare.svelte';