+++ packages/components/ui/CHANGELOG.md
@@ -1,3 +1,25 @@
-## `@svizzle/ui` v0.1.0
+## `@svizzle/ui` v0.2.0 (next)`
 
-- add `<Switch />`, a simple toggle between 2 states
+- add `src/icons/Icon.svelte`
+- add `src/icons/IconCheck.svelte`
+- add `src/icons/IconCheckSquare.svelte`
+- add `src/icons/IconChevronDown.svelte`
+- add `src/icons/IconChevronLeft.svelte`
+- add `src/icons/IconChevronRight.svelte`
+- add `src/icons/IconChevronUp.svelte`
+- add `src/icons/IconClipboard.svelte`
+- add `src/icons/IconDelete.svelte`
+- add `src/icons/IconDownload.svelte`
+- add `src/icons/IconExternalLink.svelte`
+- add `src/icons/IconFilter.svelte`
+- add `src/icons/IconGlobe.svelte`
+- add `src/icons/IconHelpCircle.svelte`
+- add `src/icons/IconInfo.svelte`
+- add `src/icons/IconSettings.svelte`
+- add `src/icons/IconSliders.svelte`
+- add `src/icons/IconSquare.svelte`
+
+
+## `@svizzle/ui` v0.1.0`
+`
+- add `src/icons/src/Switch.svelte`, a simple toggle between 2 states`